refactor: 파일 분리 및 사용하지 않는 파일 제거
