+++ android/app/src/main/AndroidManifest.xml
@@ -29,10 +29,6 @@
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".presentation.search.SearchActivity"
-            android:exported="false"
-            android:screenOrientation="portrait" />
-        <activity
             android:name=".presentation.course.CoursesActivity"
             android:exported="true"
             android:screenOrientation="portrait">
