--- android/app/src/main/java/io/coursepick/coursepick/presentation/search/OnSelectListener.kt
@@ -1,7 +1,0 @@
-package io.coursepick.coursepick.presentation.search
-
-import io.coursepick.coursepick.domain.search.Place
-
-fun interface OnSelectListener {
-    fun select(place: Place)
-}
