--- android/app/src/main/java/io/coursepick/coursepick/presentation/search/SearchAdapter.kt
@@ -1,37 +1,0 @@
-package io.coursepick.coursepick.presentation.search
-
-import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.ListAdapter
-import io.coursepick.coursepick.domain.search.Place
-
-class SearchAdapter(
-    private val onSelectListener: OnSelectListener,
-) : ListAdapter<Place, SearchViewHolder>(diffUtil) {
-    override fun onCreateViewHolder(
-        parent: ViewGroup,
-        viewType: Int,
-    ): SearchViewHolder = SearchViewHolder(parent, onSelectListener)
-
-    override fun onBindViewHolder(
-        holder: SearchViewHolder,
-        position: Int,
-    ) {
-        holder.bind(getItem(position))
-    }
-
-    companion object {
-        private val diffUtil =
-            object : DiffUtil.ItemCallback<Place>() {
-                override fun areItemsTheSame(
-                    oldItem: Place,
-                    newItem: Place,
-                ): Boolean = oldItem.id == newItem.id
-
-                override fun areContentsTheSame(
-                    oldItem: Place,
-                    newItem: Place,
-                ): Boolean = oldItem == newItem
-            }
-    }
-}
